# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/appointments": {
      "post": {
        "description": "La duración se calcula sumando los servicios, con la duración propia del profesional y el margen de limpieza de cada uno. Precio y duración quedan **congelados**: subir la tarifa mañana no altera esta cita. Con `appointments.create.own`, solo puede agendarse a sí misma.",
        "operationId": "Appointments_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScheduleAppointmentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppointmentResponse"
                }
              }
            }
          },
          "401": {
            "description": "Token ausente, inválido o caducado"
          },
          "403": {
            "description": "Con ámbito propio, el profesional de la cita no puede ser otra persona\n\nRequiere alguno de: appointments.create | appointments.create.own"
          },
          "409": {
            "description": "El profesional ya tiene otra cita en esa franja. Lo garantiza una restricción `EXCLUDE` de PostgreSQL, así que es imposible saltárselo ni con peticiones simultáneas."
          },
          "422": {
            "description": "Fuera del horario, servicio retirado o profesional que no realiza el servicio"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Agendar una cita",
        "tags": [
          "Agenda"
        ]
      },
      "get": {
        "description": "Con `appointments.read` se ve toda la agenda del salón. Con `appointments.read.own` solo la propia, y el filtro se impone en el servidor: enviar `stylistId` de otra persona no amplía el alcance.",
        "operationId": "Appointments_list",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Página a devolver, empezando en 1",
            "schema": {
              "minimum": 1,
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Elementos por página. El máximo de 100 es deliberado: sin tope, un cliente podría pedir la tabla entera y tumbar el servicio sin necesidad de un ataque.",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "example": 20,
              "type": "number"
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "description": "Ordenación como `campo:dirección`, separando por comas para ordenar por varios. Los campos admitidos dependen del recurso y se validan contra una lista blanca.",
            "schema": {
              "pattern": "^[a-zA-Z][a-zA-Z0-9_.]*:(asc|desc)(,[a-zA-Z][a-zA-Z0-9_.]*:(asc|desc))*$",
              "example": "createdAt:desc,name:asc",
              "type": "string"
            }
          },
          {
            "name": "includeDeleted",
            "required": false,
            "in": "query",
            "description": "Incluye los registros eliminados (soft delete). Requiere el permiso `*.restore` del recurso; sin él se ignora silenciosamente.",
            "schema": {
              "default": false,
              "type": "boolean"
            }
          },
          {
            "name": "stylistId",
            "required": false,
            "in": "query",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "clientId",
            "required": false,
            "in": "query",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "SCHEDULED",
                "CONFIRMED",
                "IN_PROGRESS",
                "COMPLETED",
                "CANCELLED",
                "NO_SHOW"
              ]
            }
          },
          {
            "name": "onlyBlocking",
            "required": false,
            "in": "query",
            "description": "Solo las que ocupan sillón: pendientes, confirmadas o en curso",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "from",
            "required": false,
            "in": "query",
            "schema": {
              "format": "date-time",
              "type": "string"
            }
          },
          {
            "name": "to",
            "required": false,
            "in": "query",
            "schema": {
              "format": "date-time",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/AppointmentResponse"
                      }
                    },
                    "meta": {
                      "$ref": "#/components/schemas/PageMetaResponse"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Token ausente, inválido o caducado"
          },
          "403": {
            "description": "Requiere alguno de: appointments.read | appointments.read.own"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Listar citas",
        "tags": [
          "Agenda"
        ]
      }
    },
    "/api/v1/appointments/availability": {
      "get": {
        "description": "Devuelve **momentos de inicio posibles**, no bloques libres: es lo que necesita la parrilla de horas de la interfaz. Tiene en cuenta el horario del profesional, sus ausencias, las citas ya reservadas y el margen de limpieza de cada servicio.",
        "operationId": "Appointments_availability",
        "parameters": [
          {
            "name": "stylistId",
            "required": true,
            "in": "query",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "date",
            "required": true,
            "in": "query",
            "description": "Día en la zona horaria del salón",
            "schema": {
              "pattern": "^\\d{4}-\\d{2}-\\d{2}$",
              "example": "2026-09-10",
              "type": "string"
            }
          },
          {
            "name": "serviceIds",
            "required": true,
            "in": "query",
            "description": "Servicios que se quieren reservar. Determinan cuánto tiene que durar el hueco.",
            "schema": {
              "minItems": 1,
              "maxItems": 20,
              "maxLength": 100,
              "type": "array",
              "items": {
                "type": "string",
                "format": "uuid"
              }
            }
          },
          {
            "name": "granularityMinutes",
            "required": false,
            "in": "query",
            "description": "Cada cuántos minutos se ofrece un inicio posible",
            "schema": {
              "minimum": 5,
              "maximum": 120,
              "default": 15,
              "type": "number"
            }
          },
          {
            "name": "minimumNoticeMinutes",
            "required": false,
            "in": "query",
            "description": "Antelación mínima para reservar, en minutos. Da margen a preparar el material.",
            "schema": {
              "minimum": 0,
              "maximum": 10080,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AvailableSlotResponse"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Token ausente, inválido o caducado"
          },
          "403": {
            "description": "Requiere alguno de: appointments.read | appointments.read.own"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Huecos libres para reservar",
        "tags": [
          "Agenda"
        ]
      }
    },
    "/api/v1/appointments/calendar": {
      "get": {
        "description": "Todas las citas de un rango, para pintar la parrilla. Máximo 31 días: un año entero traería decenas de miles de filas que ninguna pantalla puede mostrar.",
        "operationId": "Appointments_calendar",
        "parameters": [
          {
            "name": "from",
            "required": true,
            "in": "query",
            "schema": {
              "format": "date-time",
              "type": "string"
            }
          },
          {
            "name": "to",
            "required": true,
            "in": "query",
            "description": "Máximo 31 días desde `from`",
            "schema": {
              "format": "date-time",
              "type": "string"
            }
          },
          {
            "name": "stylistIds",
            "required": false,
            "in": "query",
            "description": "Limita la vista a estos profesionales. Sin indicar, todos.",
            "schema": {
              "maxItems": 50,
              "type": "array",
              "items": {
                "type": "string",
                "format": "uuid"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AppointmentResponse"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Token ausente, inválido o caducado"
          },
          "403": {
            "description": "Requiere alguno de: appointments.read | appointments.read.own"
          },
          "422": {
            "description": "El rango supera 31 días"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Vista de calendario",
        "tags": [
          "Agenda"
        ]
      }
    },
    "/api/v1/appointments/{id}": {
      "get": {
        "operationId": "Appointments_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppointmentResponse"
                }
              }
            }
          },
          "401": {
            "description": "Token ausente, inválido o caducado"
          },
          "403": {
            "description": "La cita es de otro profesional y solo tiene ámbito propio\n\nRequiere alguno de: appointments.read | appointments.read.own"
          },
          "404": {
            "description": "No existe o pertenece a otro salón"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Consultar una cita",
        "tags": [
          "Agenda"
        ]
      }
    },
    "/api/v1/appointments/{id}/reschedule": {
      "patch": {
        "description": "Cambiar la hora conserva la duración y **devuelve la cita a pendiente** si estaba confirmada: la clienta confirmó aquella hora, no ésta.",
        "operationId": "Appointments_reschedule",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RescheduleAppointmentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppointmentResponse"
                }
              }
            }
          },
          "401": {
            "description": "Token ausente, inválido o caducado"
          },
          "403": {
            "description": "Requiere el permiso: appointments.update"
          },
          "409": {
            "description": "La nueva franja choca con otra cita"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Mover o reasignar una cita",
        "tags": [
          "Agenda"
        ]
      }
    },
    "/api/v1/appointments/{id}/confirm": {
      "patch": {
        "operationId": "Appointments_confirm",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppointmentResponse"
                }
              }
            }
          },
          "401": {
            "description": "Token ausente, inválido o caducado"
          },
          "403": {
            "description": "Requiere el permiso: appointments.update"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Confirmar la cita con la clienta",
        "tags": [
          "Agenda"
        ]
      }
    },
    "/api/v1/appointments/{id}/start": {
      "patch": {
        "description": "Una estilista puede iniciar **sus** citas con `appointments.update.own`.",
        "operationId": "Appointments_start",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppointmentResponse"
                }
              }
            }
          },
          "401": {
            "description": "Token ausente, inválido o caducado"
          },
          "403": {
            "description": "Requiere alguno de: appointments.update | appointments.update.own"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Marcar el inicio de la cita",
        "tags": [
          "Agenda"
        ]
      }
    },
    "/api/v1/appointments/{id}/complete": {
      "patch": {
        "description": "Si nadie marcó el inicio se asume la hora prevista: en un salón con trabajo nadie pulsa «empezar», y negarse a cerrar la cita por eso solo haría que dejasen de usar el sistema.",
        "operationId": "Appointments_complete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppointmentResponse"
                }
              }
            }
          },
          "401": {
            "description": "Token ausente, inválido o caducado"
          },
          "403": {
            "description": "Requiere alguno de: appointments.update | appointments.update.own"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Completar la cita",
        "tags": [
          "Agenda"
        ]
      }
    },
    "/api/v1/appointments/{id}/cancel": {
      "patch": {
        "description": "Libera el hueco de inmediato. **Es irreversible**: si la clienta vuelve a llamar se reserva una nueva, para que el histórico conserve que hubo una cancelación.",
        "operationId": "Appointments_cancel",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CancelAppointmentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppointmentResponse"
                }
              }
            }
          },
          "401": {
            "description": "Token ausente, inválido o caducado"
          },
          "403": {
            "description": "Requiere el permiso: appointments.cancel"
          },
          "422": {
            "description": "La cita ya está terminada"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Cancelar la cita",
        "tags": [
          "Agenda"
        ]
      }
    },
...
  },
  "info": {
    "title": "AppSalonBelleza API",
...
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Desarrollo local"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Access token obtenido en POST /api/v1/auth/login"
      }
    },
    "schemas": {
...
      "ScheduleAppointmentDto": {
        "type": "object",
        "properties": {
          "clientId": {
            "type": "string",
            "format": "uuid"
          },
          "stylistId": {
            "type": "string",
            "format": "uuid"
          },
          "startsAt": {
            "format": "date-time",
            "type": "string",
            "description": "Instante de inicio en ISO 8601. La duración se calcula de los servicios.",
            "example": "2026-09-10T09:00:00.000Z"
          },
          "serviceIds": {
            "minItems": 1,
            "maxItems": 20,
            "description": "Servicios de la cita, en orden. Su precio y su duración se **congelan** al reservar: cambiar la tarifa del catálogo después no altera esta cita.",
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid",
              "maxLength": 100
            }
          },
          "source": {
            "enum": [
              "WALK_IN",
              "PHONE",
              "ONLINE",
              "STAFF"
            ],
            "type": "string",
            "default": "WALK_IN"
          },
          "notes": {
            "type": "string",
            "nullable": true,
            "maxLength": 1000,
            "description": "Notas visibles para la clienta"
          },
          "internalNotes": {
            "type": "string",
            "nullable": true,
            "maxLength": 1000,
            "description": "Notas internas del equipo: fórmulas, avisos. No se muestran a la clienta."
          },
          "force": {
            "type": "boolean",
            "default": false,
            "description": "Reserva fuera del horario del profesional. Queda registrado en la auditoría. El solapamiento con otra cita **nunca** se puede forzar: lo impide la base de datos."
          }
        },
        "required": [
          "clientId",
          "stylistId",
          "startsAt",
          "serviceIds"
        ]
      },
...
    }
  }
}